UCP/EP/FT: refactoring 2
