+++ src/ucp/wireup/wireup.c
@@ -415,7 +415,7 @@ ucp_wireup_match_p2p_lanes(ucp_ep_h ep,
     }
 }
 
-static ucs_status_t
+ucs_status_t
 ucp_wireup_find_remote_p2p_addr(ucp_ep_h ep, ucp_lane_index_t remote_lane,
                                 const ucp_unpacked_address_t *remote_address,
                                 const ucp_address_entry_t **address_entry_p,
@@ -495,7 +495,7 @@ static int ucp_ep_has_wireup_msg_pending(ucp_ep_h ucp_ep)
     return 0;
 }
 
-static void ucp_wireup_eps_progress_sched(ucp_ep_h ucp_ep)
+void ucp_wireup_eps_progress_sched(ucp_ep_h ucp_ep)
 {
     /* Use one-shot mode to avoid the need to save prog_id */
     ucs_callbackq_add_oneshot(&ucp_ep->worker->uct->progress_q, ucp_ep,
@@ -968,341 +968,7 @@ void ucp_wireup_process_ack(ucp_worker_h worker, ucp_ep_h ep,
     ucp_wireup_remote_connected(ep);
 }
 
-/*
- * --- Failed-lane recovery -------------------------------------------------
- *
- * Driven off UCP_LANE_TYPE_FAILED in cfg_key (read via
- * ucp_ep_get_failed_lanes(ep)). After ucp_ep_failover_reconfig() marks lanes
- * FAILED, it arms per-EP retry state (recovery_next_time /
- * recovery_retries_left) and lets the worker keepalive progress drive
- * subsequent ticks. Each tick calls ucp_ep_recovery_tick() which is
- * rate-limited by context->recovery_interval and bounded by
- * context->recovery_retries. A tick that is due sends one
- * UCP_WIREUP_MSG_LANES_ADDR_REQUEST on the operable AM lane after:
- *   1) replacing every failed UCT stub on ep with a wireup proxy EP,
- *   2) packing iface addresses (and p2p ep_addrs) of the failed-lane rscs.
- *
- * Peer receives the REQUEST, rebuilds its own side's UCT EPs for the lanes
- * in msg->provided_lane_map (trusting the initiator - asymmetric failure is
- * a first-class case; a still-live local UCT EP is discarded via
- * ucp_worker_discard_uct_ep), replies with its own addresses for the
- * subset it could satisfy, and clears UCP_LANE_TYPE_FAILED for just those
- * lanes. Initiator handles the REPLY symmetrically. Partial recovery is a
- * legal steady state; any remaining failed lanes are picked up by the next
- * keepalive-driven recovery tick.
- *
- * Supports both CONNECT_TO_IFACE (one-sided iface address exchange) and p2p
- * (CONNECT_TO_EP) transports. For p2p, the inner transport EP - e.g. the RC
- * QP - is created eagerly during prepare so its ep_addr can travel in the
- * LANES_ADDR_REQUEST, and the connect_to_ep handshake is finalized
- * symmetrically on both sides from the received peer ep_addr.
- *
- * The timer/retry mechanism also inherently handles the ep_count discard
- * race: the first recovery round only fires recovery_interval after
- * failover, by which time the async ucp_ep_discard_lanes() callback has
- * finalized the ep_count reactivate transition.
- */
-
-/* Ensure the lane holds a wireup proxy ready to accept a new transport EP.
- *
- * By the time we reach this function the lane is either (a) already a
- * wireup proxy (idempotent path from an earlier recovery step), or (b) a
- * failed-stub UCT EP. Case (b) only happens after ucp_ep_failover_reconfig
- * ran for the lane (either the local failover triggered by a UCT error, or
- * the explicit call the LANES_ADDR_REQUEST handler makes for asymmetric
- * failures). If a still-live transport EP shows up here it's a logic bug -
- * we should have gone through failover first.
- *
- * For p2p (CONNECT_TO_EP) lanes, also eagerly create an inner transport UCT
- * EP (e.g. RC QP) attached to the iface only. That gives us a stable local
- * ep_addr (QP number) to pack into LANES_ADDR_REQUEST / REPLY before the
- * peer's ep_addr is known. UCP_WIREUP_EP_FLAG_LOCAL_CONNECTED is not set
- * here - it is set when ucp_wireup_ep_connect_to_ep_v2() runs against the
- * peer's ep_addr. */
-static ucs_status_t
-ucp_ep_recovery_install_wireup_ep(ucp_ep_h ep, ucp_lane_index_t lane)
-{
-    const ucp_ep_config_key_t *k = &ucp_ep_config(ep)->key;
-    uct_ep_h old_uct_ep          = ucp_ep_get_lane(ep, lane);
-    uct_ep_h wireup_ep_uct, inner_uct_ep;
-    ucp_wireup_ep_t *wireup_ep;
-    uct_ep_params_t inner_params;
-    ucp_rsc_index_t rsc_index;
-    ucp_worker_iface_t *wiface;
-    ucs_status_t status;
-
-    if ((old_uct_ep != NULL) && ucp_wireup_ep_test(old_uct_ep)) {
-        return UCS_OK;
-    }
-
-    /* The only transport EP we expect here is the failed-stub placed on the
-     * lane by ucp_ep_discard_lanes() inside ucp_ep_failover_reconfig(). */
-    ucs_assertv((old_uct_ep == NULL) || ucp_is_uct_ep_failed(old_uct_ep),
-                "ep %p lane %d: unexpected live uct_ep=%p - recovery path "
-                "must be preceded by ucp_ep_failover_reconfig()",
-                ep, lane, old_uct_ep);
-
-    status = ucp_wireup_ep_create(ep, &wireup_ep_uct);
-    if (status != UCS_OK) {
-        return status;
-    }
-
-    ucs_trace("ep %p: recovery lane[%d] %p -> wireup_ep %p", ep, lane,
-              old_uct_ep, wireup_ep_uct);
-    ucp_ep_set_lane(ep, lane, wireup_ep_uct);
-
-    if (old_uct_ep != NULL) {
-        /* failed stub: decrement destroy_counter via ucp_ep_failed_destroy */
-        uct_ep_destroy(old_uct_ep);
-    }
-
-    /* For p2p transports, create the inner transport EP right away so its
-     * ep_addr (e.g. RC QP number) can be packed into LANES_ADDR_REQUEST. */
-    if (ucp_ep_is_lane_p2p(ep, lane)) {
-        rsc_index               = k->lanes[lane].rsc_index;
-        wiface                  = ucp_worker_iface(ep->worker, rsc_index);
-        inner_params.field_mask = UCT_EP_PARAM_FIELD_IFACE |
-                                  UCT_EP_PARAM_FIELD_PATH_INDEX;
-        inner_params.iface      = wiface->iface;
-        inner_params.path_index = k->lanes[lane].path_index;
-
-        status = uct_ep_create(&inner_params, &inner_uct_ep);
-        if (status != UCS_OK) {
-            ucs_diag("ep %p: failed to create inner p2p uct_ep for recovery "
-                     "lane %d: %s",
-                     ep, lane, ucs_status_string(status));
-            /* Leave the wireup proxy in place; recovery will skip this lane
-             * and retry on a future round. */
-            return status;
-        }
-
-        wireup_ep = ucp_wireup_ep(wireup_ep_uct);
-        ucs_assert(wireup_ep != NULL);
-        ucp_proxy_ep_set_uct_ep(&wireup_ep->super, inner_uct_ep, 1, rsc_index);
-
-        ucs_debug("ep %p: recovery created inner p2p uct_ep=%p for lane %d"
-                  " (rsc=%d path=%u)",
-                  ep, inner_uct_ep, lane, rsc_index,
-                  k->lanes[lane].path_index);
-    }
-
-    return UCS_OK;
-}
-
-/* Attach a freshly-created real UCT EP as next_ep of the lane's wireup proxy
- * and mark it ready + remote-connected so that ucp_wireup_eps_progress swaps
- * it out on the next scheduled run. */
-static void
-ucp_ep_recovery_attach_next_ep(ucp_ep_h ep, ucp_lane_index_t lane,
-                               uct_ep_h next_ep)
-{
-    uct_ep_h proxy             = ucp_ep_get_lane(ep, lane);
-    ucp_wireup_ep_t *wireup_ep = ucp_wireup_ep(proxy);
-
-    ucs_assert(wireup_ep != NULL);
-    ucs_assert(wireup_ep->super.uct_ep == NULL);
-
-    ucp_wireup_ep_set_next_ep(proxy, next_ep, ucp_ep_get_rsc_index(ep, lane));
-    wireup_ep->flags |= UCP_WIREUP_EP_FLAG_READY |
-                        UCP_WIREUP_EP_FLAG_REMOTE_CONNECTED;
-}
-
-/* Mark an already-attached p2p wireup proxy as ready+remote-connected, after
- * ucp_wireup_ep_connect_to_ep_v2() has successfully connected its inner EP. */
-static void
-ucp_ep_recovery_mark_p2p_ready(ucp_ep_h ep, ucp_lane_index_t lane)
-{
-    uct_ep_h proxy             = ucp_ep_get_lane(ep, lane);
-    ucp_wireup_ep_t *wireup_ep = ucp_wireup_ep(proxy);
-
-    ucs_assert(wireup_ep != NULL);
-    ucs_assert(wireup_ep->super.uct_ep != NULL);
-
-    wireup_ep->flags |= UCP_WIREUP_EP_FLAG_READY |
-                        UCP_WIREUP_EP_FLAG_REMOTE_CONNECTED;
-}
-
-/* Find a CONNECT_TO_IFACE address entry whose TL matches the one used
- * locally by `lane`. For p2p transports use ucp_wireup_find_remote_p2p_addr()
- * which also drills into a matching ep_addr. */
-static const ucp_address_entry_t *
-ucp_ep_recovery_find_iface_addr(ucp_ep_h ep, ucp_lane_index_t lane,
-                                const ucp_unpacked_address_t *remote_address)
-{
-    ucp_context_h context     = ep->worker->context;
-    ucp_rsc_index_t rsc_index = ucp_ep_config(ep)->key.lanes[lane].rsc_index;
-    uint16_t tl_name_csum     = context->tl_rscs[rsc_index].tl_name_csum;
-    const ucp_address_entry_t *ae;
-
-    ucp_unpacked_address_for_each(ae, remote_address) {
-        if ((ae->tl_name_csum == tl_name_csum) && (ae->iface_addr != NULL)) {
-            return ae;
-        }
-    }
-
-    return NULL;
-}
-
-/* Rebuild one CONNECT_TO_IFACE lane: create a UCT EP to the remote iface and
- * attach it as next_ep of the lane's wireup proxy. */
-static ucs_status_t
-ucp_ep_recovery_rebuild_iface_lane(
-        ucp_ep_h ep, ucp_lane_index_t lane,
-        const ucp_unpacked_address_t *remote_address)
-{
-    const ucp_ep_config_key_t *k = &ucp_ep_config(ep)->key;
-    ucp_worker_h worker          = ep->worker;
-    const ucp_address_entry_t *ae;
-    uct_ep_params_t uct_ep_params;
-    ucp_worker_iface_t *wiface;
-    ucp_rsc_index_t rsc_index;
-    uct_ep_h uct_ep;
-    ucs_status_t status;
-
-    ae = ucp_ep_recovery_find_iface_addr(ep, lane, remote_address);
-    if (ae == NULL) {
-        ucs_debug("ep %p: no remote iface address for lane %d", ep, lane);
-        return UCS_ERR_UNREACHABLE;
-    }
-
-    status = ucp_ep_recovery_install_wireup_ep(ep, lane);
-    if (status != UCS_OK) {
-        return status;
-    }
-
-    rsc_index                = k->lanes[lane].rsc_index;
-    wiface                   = ucp_worker_iface(worker, rsc_index);
-    uct_ep_params.field_mask = UCT_EP_PARAM_FIELD_IFACE      |
-                               UCT_EP_PARAM_FIELD_DEV_ADDR   |
-                               UCT_EP_PARAM_FIELD_IFACE_ADDR |
-                               UCT_EP_PARAM_FIELD_PATH_INDEX;
-    uct_ep_params.iface      = wiface->iface;
-    uct_ep_params.dev_addr   = ae->dev_addr;
-    uct_ep_params.iface_addr = ae->iface_addr;
-    uct_ep_params.path_index = k->lanes[lane].path_index;
-
-    status = uct_ep_create(&uct_ep_params, &uct_ep);
-    if (status != UCS_OK) {
-        ucs_diag("ep %p: uct_ep_create failed for recovery lane %d: %s", ep,
-                 lane, ucs_status_string(status));
-        return status;
-    }
-
-    ucp_ep_recovery_attach_next_ep(ep, lane, uct_ep);
-    ucs_debug("ep %p: recovered iface-lane[%d]=%p via rsc[%d]", ep, lane,
-              uct_ep, rsc_index);
-    return UCS_OK;
-}
-
-/* Rebuild one p2p (CONNECT_TO_EP) lane: ensure the wireup proxy holds an
- * inner transport EP (created in prepare), find the remote ep_addr for this
- * lane, and connect via ucp_wireup_ep_connect_to_ep_v2(). */
-static ucs_status_t
-ucp_ep_recovery_rebuild_p2p_lane(
-        ucp_ep_h ep, ucp_lane_index_t lane,
-        const ucp_unpacked_address_t *remote_address)
-{
-    const ucp_address_entry_t *address_entry;
-    const ucp_address_entry_ep_addr_t *ep_entry;
-    ucp_lane_index_t remote_lane;
-    ucs_status_t status;
-
-    /* Idempotent: if lane already has a wireup proxy with an inner p2p EP,
-     * this is a no-op. Otherwise (peer side receiving REQUEST while its own
-     * lane still holds a failed stub), we create the inner EP here. */
-    status = ucp_ep_recovery_install_wireup_ep(ep, lane);
-    if (status != UCS_OK) {
-        return status;
-    }
-
-    /* Symmetric assumption for this POC: the peer's lane index mirrors ours.
-     * ucp_address_pack() with lanes2remote==NULL stamps the sender-side local
-     * lane as the remote_lane tag in the ep_addr, so we look up by our own
-     * lane index. */
-    remote_lane = lane;
-    status = ucp_wireup_find_remote_p2p_addr(ep, remote_lane, remote_address,
-                                             &address_entry, &ep_entry);
-    if (status != UCS_OK) {
-        ucs_debug("ep %p: no remote ep_addr for p2p lane %d", ep, lane);
-        return status;
-    }
-
-    status = ucp_wireup_ep_connect_to_ep_v2(ucp_ep_get_lane(ep, lane),
-                                            address_entry, ep_entry);
-    if (status != UCS_OK) {
-        ucs_diag("ep %p: connect_to_ep_v2 failed for recovery p2p lane %d: %s",
-                 ep, lane, ucs_status_string(status));
-        return status;
-    }
-
-    ucp_ep_recovery_mark_p2p_ready(ep, lane);
-    ucs_debug("ep %p: recovered p2p-lane[%d] via rsc[%d]", ep, lane,
-              ucp_ep_get_rsc_index(ep, lane));
-    return UCS_OK;
-}
-
-/* For each lane in `lanes_to_rebuild`, bring its UCT resources back online.
- * Dispatches to the iface-connect or p2p helper based on lane type. Returns
- * the bitmap of lanes successfully rebuilt; lanes not in the returned bitmap
- * stay UCP_LANE_TYPE_FAILED and will be retried on a future recovery round. */
-static ucp_lane_map_t
-ucp_ep_recovery_rebuild_lanes(ucp_ep_h ep, ucp_lane_map_t lanes_to_rebuild,
-                              const ucp_unpacked_address_t *remote_address)
-{
-    const ucp_ep_config_key_t *k = &ucp_ep_config(ep)->key;
-    ucp_lane_map_t rebuilt       = 0;
-    ucp_lane_index_t lane;
-    ucs_status_t status;
-
-    ucs_for_each_bit(lane, lanes_to_rebuild) {
-        if (lane >= k->num_lanes) {
-            continue;
-        }
-
-        if (ucp_ep_is_lane_p2p(ep, lane)) {
-            status = ucp_ep_recovery_rebuild_p2p_lane(ep, lane,
-                                                      remote_address);
-        } else {
-            status = ucp_ep_recovery_rebuild_iface_lane(ep, lane,
-                                                        remote_address);
-        }
-
-        if (status == UCS_OK) {
-            rebuilt |= UCS_BIT(lane);
-        }
-    }
-
-    return rebuilt;
-}
-
-/* Replace failed UCT stubs with wireup proxies for the given lane set. For
- * p2p lanes this also creates the inner transport EP so its ep_addr is
- * available at packing time. Returns the bitmap of lanes that now hold a
- * wireup proxy in the requested state. */
-static ucp_lane_map_t
-ucp_ep_recovery_prepare_lanes(ucp_ep_h ep, ucp_lane_map_t lanes)
-{
-    const ucp_ep_config_key_t *cfg_key = &ucp_ep_config(ep)->key;
-    ucp_lane_map_t lane_map            = 0;
-    ucp_lane_index_t lane;
-
-    ucs_for_each_bit(lane, lanes) {
-        if (lane >= cfg_key->num_lanes) {
-            continue;
-        }
-
-        if (ucp_ep_recovery_install_wireup_ep(ep, lane) != UCS_OK) {
-            continue;
-        }
-
-        lane_map |= UCS_BIT(lane);
-    }
-
-    return lane_map;
-}
-
-static void
-ucp_ep_recovery_send_lanes_addr_msg(ucp_ep_h ep, uint8_t msg_type,
+void ucp_wireup_send_lanes_addr_msg(ucp_ep_h ep, uint8_t msg_type,
                                     ucp_lane_map_t requested_lane_map,
                                     ucp_lane_map_t provided_lane_map)
 {
@@ -1321,199 +987,6 @@ ucp_ep_recovery_send_lanes_addr_msg(ucp_ep_h ep, uint8_t msg_type,
         ucs_diag("ep %p: failed to send %s: %s", ep,
                  ucp_wireup_msg_str(msg_type), ucs_status_string(status));
     }
-}
-
-/* Send one LANES_ADDR_REQUEST for the current set of failed lanes. Idempotent
- * across retries: proxies already in place are left alone, only missing ones
- * are installed. */
-static void ucp_ep_recovery_send_request(ucp_ep_h ep)
-{
-    ucp_lane_map_t failed_lanes = ucp_ep_get_failed_lanes(ep);
-    ucp_lane_map_t recovery_lanes;
-
-    ucs_assert(ucp_ep_config(ep)->key.am_lane != UCP_NULL_LANE);
-
-    recovery_lanes = ucp_ep_recovery_prepare_lanes(ep, failed_lanes);
-    if (recovery_lanes == 0) {
-        ucs_diag("ep %p: no lanes to recover", ep);
-        return;
-    }
-
-    ucs_debug("ep %p: sending recovery request, failed=0x%" PRIx64
-              " recovery=0x%" PRIx64, ep, failed_lanes, recovery_lanes);
-
-    ucp_ep_recovery_send_lanes_addr_msg(ep, UCP_WIREUP_MSG_LANES_ADDR_REQUEST,
-                                        failed_lanes, recovery_lanes);
-}
-
-/* A lane whose UCT EP is either a READY wireup proxy (about to be swapped
- * by ucp_wireup_eps_progress) or a real transport EP (already swapped) is
- * effectively recovered - its FAILED bit can be cleared. A failed stub or
- * a wireup proxy without READY still needs work. */
-static int ucp_ep_recovery_lane_is_ready(ucp_ep_h ep, ucp_lane_index_t lane)
-{
-    uct_ep_h uct_ep = ucp_ep_get_lane(ep, lane);
-    ucp_wireup_ep_t *wireup_ep;
-
-    if (uct_ep == NULL) {
-        return 0;
-    }
-
-    if (ucp_is_uct_ep_failed(uct_ep)) {
-        return 0;
-    }
-
-    wireup_ep = ucp_wireup_ep(uct_ep);
-    if (wireup_ep == NULL) {
-        /* ucp_wireup_eps_progress already replaced the proxy with the real
-         * transport EP - lane is fully operational. */
-        return 1;
-    }
-
-    return !!(wireup_ep->flags & UCP_WIREUP_EP_FLAG_READY);
-}
-
-static ucp_lane_map_t ucp_ep_recovery_get_ready_lanes(ucp_ep_h ep,
-                                                      ucp_lane_map_t failed_lanes)
-{
-    ucp_lane_map_t ready_lanes = 0;
-    ucp_lane_index_t lane;
-
-    ucs_for_each_bit(lane, failed_lanes) {
-        if (ucp_ep_recovery_lane_is_ready(ep, lane)) {
-            ready_lanes |= UCS_BIT(lane);
-        }
-    }
-
-    return ready_lanes;
-}
-
-/* Heap-resident retry state for failed-lane recovery. Allocated by
- * ucp_ep_recovery_arm() on failover, kept alive by the callbackq while the
- * oneshot is pending, freed either by the oneshot itself on completion / on
- * retry exhaustion, or by ucp_ep_recovery_remove_filter() during endpoint
- * teardown. No per-EP state is kept on ucp_ep_t / ucp_ep_ext_t. */
-typedef struct ucp_ep_recovery_arg {
-    ucp_ep_h       ucp_ep;        /* Back-pointer used for callbackq key and
-                                     filter matching. */
-    ucs_time_t     next_time;     /* Earliest time at which the next round
-                                     may send LANES_ADDR_REQUEST. */
-    unsigned       retries_left;  /* Rounds remaining before the endpoint is
-                                     declared fully failed. */
-} ucp_ep_recovery_arg_t;
-
-/* Enqueue a recovery oneshot for @a arg. Caller keeps ownership of @a arg
- * until the oneshot runs (it may free itself) or is purged via the filter. */
-static void ucp_ep_recovery_arg_enqueue(ucp_ep_recovery_arg_t *arg)
-{
-    ucp_worker_h worker = arg->ucp_ep->worker;
-
-    ucs_callbackq_add_oneshot(&worker->uct->progress_q, arg->ucp_ep,
-                              ucp_ep_recovery_progress, arg);
-    ucp_worker_signal_internal(worker);
-}
-
-int ucp_ep_recovery_remove_filter(const ucs_callbackq_elem_t *elem, void *arg)
-{
-    ucp_ep_recovery_arg_t *recovery_arg = elem->arg;
-
-    if ((elem->cb == ucp_ep_recovery_progress) &&
-        (recovery_arg->ucp_ep == arg)) {
-        ucs_free(recovery_arg);
-        return 1;
-    }
-
-    return 0;
-}
-
-ucs_status_t ucp_ep_recovery_arm(ucp_ep_h ep)
-{
-    ucp_worker_h worker   = ep->worker;
-    ucp_context_h context = worker->context;
-    ucp_ep_recovery_arg_t *arg;
-
-    /* Remove (and free) any previously-queued recovery oneshot for this EP. */
-    ucs_callbackq_remove_oneshot(&worker->uct->progress_q, ep,
-                                 ucp_ep_recovery_remove_filter, ep);
-
-    arg = ucs_malloc(sizeof(*arg), "ucp_ep_recovery_arg");
-    if (arg == NULL) {
-        ucs_error("ep %p: failed to allocate recovery argument", ep);
-        return UCS_ERR_NO_MEMORY;
-    }
-
-    arg->ucp_ep       = ep;
-    arg->next_time    = ucs_get_time() + context->config.ext.recovery_interval;
-    arg->retries_left = context->config.ext.recovery_retries;
-
-    ucp_ep_recovery_arg_enqueue(arg);
-    return UCS_OK;
-}
-
-unsigned ucp_ep_recovery_progress(void *arg)
-{
-    ucp_ep_recovery_arg_t *recovery_arg = arg;
-    ucp_ep_h ep                         = recovery_arg->ucp_ep;
-    ucp_worker_h worker                 = ep->worker;
-    ucp_context_h context               = worker->context;
-    ucp_lane_map_t failed, recovered;
-    ucs_time_t now;
-    ucs_status_t status;
-
-    UCS_ASYNC_BLOCK(&worker->async);
-
-    if (ep->flags & UCP_EP_FLAG_FAILED) {
-        /* Endpoint was declared fully failed elsewhere; nothing more to do. */
-        goto done;
-    }
-
-    failed = ucp_ep_get_failed_lanes(ep);
-    if (failed == 0) {
-        /* Recovery completed between rounds. */
-        goto done;
-    }
-
-    recovered = ucp_ep_recovery_get_ready_lanes(ep, failed);
-    status    = ucp_ep_reconfig_clear_failed_lanes(ep, recovered);
-    if (status != UCS_OK) {
-        ucs_error("ep %p: failed to clear FAILED states for lanes 0x%" PRIx64,
-                  ep, recovered);
-        ucp_ep_set_lanes_failed_schedule(ep, 0, status);
-        goto done;
-    }
-
-    failed &= ~recovered;
-    if (failed == 0) {
-        goto done;
-    }
-
-    now = ucs_get_time();
-    if (now < recovery_arg->next_time) {
-        ucp_ep_recovery_arg_enqueue(recovery_arg);
-        goto out;
-    }
-
-    if (recovery_arg->retries_left == 0) {
-        ucs_error("ep %p: recovery retries exhausted", ep);
-        ucp_ep_set_lanes_failed_schedule(ep, 0, UCS_ERR_ENDPOINT_TIMEOUT);
-        goto done;
-    }
-
-    ucs_debug("ep %p: recovery round (retries_left=%u, failed=0x%" PRIx64 ")",
-              ep, recovery_arg->retries_left, (uint64_t)failed);
-
-    ucp_ep_recovery_send_request(ep);
-
-    --recovery_arg->retries_left;
-    recovery_arg->next_time = now + context->config.ext.recovery_interval;
-    ucp_ep_recovery_arg_enqueue(recovery_arg);
-    goto out;
-
-done:
-    ucs_free(recovery_arg);
-out:
-    UCS_ASYNC_UNBLOCK(&worker->async);
-    return 1;
 }
 
 static UCS_F_NOINLINE void
@@ -1568,9 +1041,8 @@ ucp_wireup_process_lanes_addr_request(
     /* Always reply, even if peer_provided == 0, so the initiator knows the
      * peer handled the request (empty provided_lane_map means "I couldn't
      * satisfy any of the lanes you asked about"). */
-    ucp_ep_recovery_send_lanes_addr_msg(ep, UCP_WIREUP_MSG_LANES_ADDR_REPLY,
-                                        msg->requested_lane_map,
-                                        peer_provided);
+    ucp_wireup_send_lanes_addr_msg(ep, UCP_WIREUP_MSG_LANES_ADDR_REPLY,
+                                   msg->requested_lane_map, peer_provided);
 
     /* The recovery tick clears UCP_LANE_TYPE_FAILED once the rebuilt lanes
      * reach a READY state (either as READY wireup proxies or as real
